tshirt_automation/requirements.txt
# requirements.txt – Python dependencies
Flask
openai
requests
pillow
